add menu home. add file utils
