+++ app/src/main/java/ru/gb/veber/kotlinmvvm/view/MainActivity.kt
@@ -30,6 +30,7 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
         menuInflater.inflate(R.menu.menu_main, menu)
+
         return super.onCreateOptionsMenu(menu)
     }
