+++ app/src/main/java/ru/gb/veber/kotlinmvvm/view/fragment/ContentProviderFragment.kt
@@ -9,21 +9,24 @@ import android.database.Cursor
 import android.net.Uri
 import android.os.Bundle
 import android.os.Handler
-import android.os.HandlerThread
 import android.provider.ContactsContract
 import android.util.Log
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import android.view.*
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
+import ru.gb.veber.kotlinmvvm.R
 import ru.gb.veber.kotlinmvvm.databinding.FragmentContentProviderBinding
+import ru.gb.veber.kotlinmvvm.view.MainActivity
 import ru.gb.veber.kotlinmvvm.view.adapter.AdapterContacts
-import ru.gb.veber.kotlinmvvm.view.adapter.CitysFragmentAdapter
 import ru.gb.veber.kotlinmvvm.view.adapter.OnContactsClickListener
 
 
 const val REQUEST_CODE = 44
+private const val CONTACT_ID = ContactsContract.Contacts._ID
+private const val DISPLAY_NAME = ContactsContract.Contacts.DISPLAY_NAME
+private const val HAS_PHONE_NUMBER = ContactsContract.Contacts.HAS_PHONE_NUMBER
+private const val PHONE_NUMBER = ContactsContract.CommonDataKinds.Phone.NUMBER
+private const val PHONE_CONTACT_ID = ContactsContract.CommonDataKinds.Phone.CONTACT_ID
 
 class ContentProviderFragment : Fragment(), OnContactsClickListener {
     private var _binding: FragmentContentProviderBinding? = null
@@ -41,21 +44,17 @@ class ContentProviderFragment : Fragment(), OnContactsClickListener {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        setHasOptionsMenu(true)
+        (activity as MainActivity).supportActionBar?.setDisplayHomeAsUpEnabled(true)
 
         binding.listContacts.adapter = adapter
         adapter.setOnContactsClickListener(this)
         checkPermission()
     }
 
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
-    companion object {
-        @JvmStatic
-        fun newInstance() =
-            ContentProviderFragment()
+    override fun onCityClick(contacts: MyContacts) {
+        val intent = Intent(Intent.ACTION_DIAL, Uri.fromParts("tel", contacts.phone, null))
+        startActivity(intent)
     }
 
     private fun checkPermission() {
@@ -116,6 +115,7 @@ class ContentProviderFragment : Fragment(), OnContactsClickListener {
         }
     }
 
+    @SuppressLint("Range")
     private fun getContacts() {
 
         var listContacts: MutableList<MyContacts> = mutableListOf()
@@ -139,7 +139,7 @@ class ContentProviderFragment : Fragment(), OnContactsClickListener {
                     null,
                     null,
                     null,
-                    ContactsContract.Contacts.DISPLAY_NAME_PRIMARY
+                    ContactsContract.Contacts.DISPLAY_NAME + " ASC"
                 )
 
                 cursorWithContacts?.let { cursor ->
@@ -152,6 +152,7 @@ class ContentProviderFragment : Fragment(), OnContactsClickListener {
                             val name = cursor.getString(pos)
                             val name2 = cursorWithContacts2?.getString(pos2!!)
                             listContacts.add(MyContacts(name, name2))
+                            Log.d("TAG", "getContacts() called with: name2 = $name2")
                         }
                     }
                 }
@@ -164,9 +165,32 @@ class ContentProviderFragment : Fragment(), OnContactsClickListener {
         }.start()
     }
 
-    override fun onCityClick(contacts: MyContacts) {
-        val intent = Intent(Intent.ACTION_DIAL, Uri.fromParts("tel", contacts.phone, null))
-        startActivity(intent)
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+
+        menu.findItem(R.id.menu_item_update).isVisible = false
+        menu.findItem(R.id.menu_content_provider).isVisible = false
+        super.onCreateOptionsMenu(menu, inflater)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            android.R.id.home -> {
+                activity?.supportFragmentManager?.popBackStack()
+                (activity as MainActivity).supportActionBar?.setDisplayHomeAsUpEnabled(false)
+            }
+        }
+        return super.onOptionsItemSelected(item)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    companion object {
+        @JvmStatic
+        fun newInstance() =
+            ContentProviderFragment()
     }
 }
